Add Functionalities, Correct Output, Adapt old func to new symbol table
Added subprogram control and control flow functionality. Modified symbol table to make it possible and corrected output by moving to show function and writing to file instead of cout

+++ JazInterpreter/JazInterpreter/ControlManipulation.h
@@ -4,7 +4,7 @@
 
 using namespace std;
 
-static size_t programCounter = 0;
+static int programCounter = 0;
 
 void separateInstructions(vector<string> instructions);
